requirements.txt
pydantic>=2.0.0; python_version < "3.13.0"
pyyaml; python_version < "3.13.0"
Pillow>=10.0.0; python_version < "3.13.0"
numpy>=1.23.0; python_version < "3.13.0"
numba>=0.50.0; python_version < "3.13.0"
torch; python_version < "3.13.0"
torchvision; python_version < "3.13.0"
lightning; python_version < "3.13.0"
einops; python_version < "3.13.0"
